Fix update_or_create usage for Deals and separate mocks from the test file
